+++ front/src/queries/mutations.ts
@@ -1,5 +1,5 @@
+import { SHIPMENT_FIELDS_FRAGMENT } from "../../../graphql/fragments";
 import { graphql } from "../../../graphql/graphql"
-import { SHIPMENT_FIELDS_FRAGMENT } from "./fragments";
 
 export const UPDATE_SHIPMENT_WHEN_RECEIVING = graphql(`
   mutation UpdateShipmentWhenReceiving(
