+++ front/src/views/Box/components/AssignBoxToShipment.tsx
@@ -3,8 +3,7 @@ import { Text, FormControl, FormErrorMessage, Button, Flex, chakra } from "@chak
 import { Select } from "chakra-react-select";
 import { IDropdownOption } from "components/Form/SelectField";
 import { ShipmentOption } from "components/Form/ShipmentOption";
-import { BoxByLabelIdentifier } from "../../../../../graphql/types";
-import { UpdateBoxMutation } from "queries/types";
+import { BoxByLabelIdentifier, UpdateBoxMutation } from "queries/types";
 
 export interface IAssignBoxToShipmentProps {
   boxData: BoxByLabelIdentifier | UpdateBoxMutation;
