+++ front/src/views/Box/components/BoxCard.tsx
@@ -28,8 +28,7 @@ import { NavLink } from "react-router-dom";
 import { colorIsBright } from "utils/helpers";
 import { Style } from "victory";
 import HistoryEntries from "./HistoryEntries";
-import { BoxByLabelIdentifier } from "../../../../../graphql/types";
-import { UpdateBoxMutation } from "queries/types";
+import { BoxByLabelIdentifier, UpdateBoxMutation } from "queries/types";
 
 export interface IBoxCardProps {
   boxData: BoxByLabelIdentifier | UpdateBoxMutation;
