+++ front/src/views/Box/components/BoxMoveLocation.tsx
@@ -1,6 +1,5 @@
 import { List, Text, Flex, WrapItem, Button, SkeletonText } from "@chakra-ui/react";
-import { BoxByLabelIdentifier } from "../../../../../graphql/types";
-import { UpdateBoxMutation } from "queries/types";
+import { BoxByLabelIdentifier, UpdateBoxMutation } from "queries/types";
 
 export interface IBoxMoveLocationProps {
   boxData: BoxByLabelIdentifier | UpdateBoxMutation;
