+++ front/src/views/Box/components/BoxTabs.tsx
@@ -21,8 +21,7 @@ import { FaDollyFlatbed, FaWarehouse } from "react-icons/fa";
 import AssignBoxToShipment from "./AssignBoxToShipment";
 
 import BoxMoveLocation from "./BoxMoveLocation";
-import { BoxByLabelIdentifier } from "../../../../../graphql/types";
-import { UpdateBoxMutation } from "queries/types";
+import { BoxByLabelIdentifier, UpdateBoxMutation } from "queries/types";
 
 export interface IBoxTabsProps {
   boxData: BoxByLabelIdentifier | UpdateBoxMutation;
