+++ front/src/views/Transfers/ShipmentView/ShipmentView.tsx
@@ -18,7 +18,6 @@ import { useCallback, useContext, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { useErrorHandling } from "hooks/useErrorHandling";
 import { useNotification } from "hooks/useNotification";
-import { SHIPMENT_FIELDS_FRAGMENT } from "queries/fragments";
 import { GlobalPreferencesContext } from "providers/GlobalPreferencesProvider";
 import { ButtonSkeleton, ShipmentCardSkeleton, TabsSkeleton } from "components/Skeletons";
 import { BoxReconciliationOverlay } from "components/BoxReconciliationOverlay/BoxReconciliationOverlay";
@@ -34,6 +33,7 @@ import ShipmentReceivingContent from "./components/ShipmentReceivingContent";
 import ShipmentReceivingCard from "./components/ShipmentReceivingCard";
 import { useLoadAndSetGlobalPreferences } from "hooks/useLoadAndSetGlobalPreferences";
 import { ShipmentState, ShipmentDetail, User } from "../../../../../graphql/types";
+import { SHIPMENT_FIELDS_FRAGMENT } from "../../../../../graphql/fragments";
 
 enum ShipmentActionEvent {
   ShipmentStarted = "Shipment Started",
